Added the anagram of my name to the file GCI-STUDENTS.txt
In this very pull request i have added the anagram to my name to the file. This is done in respect of the task listed at google code in competition. My actual name is Rohan Garg and the anagram to it is arrha gong and so as said in the task it was added alphabetically to this very file.